Clean-up current tests
