+++ idg_test.go
@@ -1,10 +1,8 @@
 package idg
 
 import (
-	"encoding/json"
 	"fmt"
 	"testing"
-	"time"
 
 	"github.com/missionMeteora/uuid"
 )
@@ -14,28 +12,7 @@ var (
 	uuidSink uuid.UUID
 )
 
-func TestTime(t *testing.T) {
-	var (
-		tt  time.Time
-		err error
-	)
-
-	// Get a current timestamp
-	now := time.Now()
-	// Ensure our new ID is at least one millisecond behind our timestamp
-	time.Sleep(time.Millisecond)
-	// Generate a new ID
-	id := newID(0)
-	// Get the time from our ID
-	if tt, err = id.Time(); err != nil {
-		t.Fatal(err)
-	}
-	// Check to see if our id's time is after our timestamp (it should be)
-	if !tt.After(now) {
-		t.Fatalf("invalid time, should be after initial timestamp: %v / %v", now, tt)
-	}
-}
-func TestIndexing(t *testing.T) {
+func TestIDGIndexing(t *testing.T) {
 	var err error
 	// Generate new ID with an index starting at 3
 	idg := New(3)
@@ -52,50 +29,6 @@ func TestIndexing(t *testing.T) {
 	}
 }
 
-func TestParse(t *testing.T) {
-	var (
-		nid ID
-		err error
-	)
-
-	// Generate an ID with the index starting at 1337
-	id := newID(1337)
-	// Get the string representation of our ID
-	sid := id.String()
-	// Parse the string to a new ID
-	if nid, err = Parse(sid); err != nil {
-		t.Fatal(err)
-	}
-	// Check if the ID's match
-	if id != nid {
-		t.Fatalf("ID's do not match: %v / %v", id.Bytes(), nid.Bytes())
-	}
-}
-
-func TestJSON(t *testing.T) {
-	var (
-		b   []byte
-		err error
-	)
-
-	// Generate an ID with the index starting at 1337
-	id := newID(1337)
-	// Marshal ID as JSON
-	if b, err = json.Marshal(&id); err != nil {
-		t.Fatal(err)
-	}
-
-	var nid ID
-	// Parse as JSON to a new ID
-	if err = json.Unmarshal(b, &nid); err != nil {
-		t.Fatal(err)
-	}
-	// Check if the ID's match
-	if id != nid {
-		t.Fatalf("ID's do not match: %v / %v", id.Bytes(), nid.Bytes())
-	}
-}
-
 func testIndex(id ID, expected uint64) (err error) {
 	var idx uint64
 	if idx, err = id.Index(); err != nil {
@@ -107,7 +40,7 @@ func testIndex(id ID, expected uint64) (err error) {
 	return
 }
 
-func BenchmarkGenerationIDG(b *testing.B) {
+func BenchmarkIDG_Gen(b *testing.B) {
 	idg := New(0)
 	for i := 0; i < b.N; i++ {
 		idSink = idg.Next()
@@ -116,7 +49,7 @@ func BenchmarkGenerationIDG(b *testing.B) {
 	b.ReportAllocs()
 }
 
-func BenchmarkGenerationUUID(b *testing.B) {
+func BenchmarkUUID_Gen(b *testing.B) {
 	ug := uuid.NewGen()
 	for i := 0; i < b.N; i++ {
 		uuidSink = ug.New()
@@ -125,7 +58,7 @@ func BenchmarkGenerationUUID(b *testing.B) {
 	b.ReportAllocs()
 }
 
-func BenchmarkGenerationParallelIDG(b *testing.B) {
+func BenchmarkIDG_Gen_Para(b *testing.B) {
 	idg := New(0)
 	b.RunParallel(func(pb *testing.PB) {
 		for pb.Next() {
@@ -136,7 +69,7 @@ func BenchmarkGenerationParallelIDG(b *testing.B) {
 	b.ReportAllocs()
 }
 
-func BenchmarkGenerationParallelUUID(b *testing.B) {
+func BenchmarkUUID_Gen_Para(b *testing.B) {
 	ug := uuid.NewGen()
 	b.RunParallel(func(pb *testing.PB) {
 		for pb.Next() {
